+++ projects/epgu-constructor/src/lib/component/custom-screen/component-list-resolver/component-list-resolver.component.ts
@@ -37,8 +37,6 @@ export class ComponentListResolverComponent implements AfterViewInit, OnChanges 
   @Input() componentsGroupIndex = 0;
   @Input() componentType: ComponentTypes;
   @Input() limit?: number | string;
-  @Input() parentFirst?: boolean;
-  @Input() parentLast?: boolean;
   componentRef: ComponentRef<ScreenComponentTypes>;
 
   constructor(
@@ -49,19 +47,10 @@ export class ComponentListResolverComponent implements AfterViewInit, OnChanges 
   ) {}
 
   ngOnChanges(changes: SimpleChanges): void {
-    if (
-      this.componentRef &&
-      (changes.componentsGroupIndex ||
-        changes.componentIndex ||
-        changes.limit ||
-        changes.parentFirst ||
-        changes.parentLast)
-    ) {
+    if (this.componentRef && (changes.componentsGroupIndex || changes.componentIndex)) {
       this.componentRef.instance.componentIndex = this.componentIndex;
       this.componentRef.instance.componentsGroupIndex = this.componentsGroupIndex;
       this.componentRef.instance.limit = this.limit;
-      this.componentRef.instance.parentFirst = this.parentFirst;
-      this.componentRef.instance.parentLast = this.parentLast;
     }
   }
 
@@ -105,8 +94,6 @@ export class ComponentListResolverComponent implements AfterViewInit, OnChanges 
     // @ts-ignore
     this.componentRef.instance.componentsGroupIndex = this.componentsGroupIndex;
     this.componentRef.instance.limit = this.limit;
-    this.componentRef.instance.parentFirst = this.parentFirst;
-    this.componentRef.instance.parentLast = this.parentLast;
   }
 
   private getComponentByType(cmpType: ComponentTypes): Type<ScreenComponentTypes> {