+++ projects/epgu-constructor/src/lib/component/unique-screen/components/select-map-object/select-map-object.component.ts
@@ -71,6 +71,7 @@ import { ActionToolsService } from '../../../../shared/directives/action/action-
 import { PriorityItemsService } from './services/priority-items/priority-items.service';
 import { ComponentValue } from '../../../../shared/services/dictionary/dictionary.interface';
 import { Invite, InviteService } from '../../../../core/services/invite/invite.service';
+import { ActivatedRoute } from '@angular/router';
 
 const INTERNAL_ERROR_MESSAGE = 'Internal Error';
 
@@ -128,6 +129,7 @@ export class SelectMapObjectComponent implements OnInit, AfterViewChecked, OnDes
     private yaMapService: YaMapService,
     private zone: NgZone,
     private priorityItemsService: PriorityItemsService,
+    private route: ActivatedRoute,
   ) {}
 
   ngOnInit(): void {
@@ -448,7 +450,10 @@ export class SelectMapObjectComponent implements OnInit, AfterViewChecked, OnDes
     }
 
     const getFilter = this.data.attrs?.isInvite
-      ? this.invite.getInvite(String(this.screenService.orderId))
+      ? this.invite.getInvite(
+          this.route.snapshot.queryParamMap.get('parentOrderId') ??
+            String(this.screenService.orderId),
+        )
       : of({} as Invite);
 
     return getFilter.pipe(