+++ projects/epgu-constructor/src/lib/shared/components/multiple-choice-dictionary/multiple-choice-dictionary/multiple-choice-dictionary.component.ts
@@ -75,7 +75,7 @@ export class MultipleChoiceDictionaryComponent implements OnInit, ControlValueAc
           if (items instanceof Error) {
             return this.openErrorModal();
           }
-          return of(items || this.selectedItems.list);
+          return of(items || this.selectedItems);
         }),
       )
       .subscribe((items) => {